requirement.txt
flask
python-dotenv
requests
textblob
wolframalpha
